data-generator/requirements.txt
# Data Generator Dependencies

# Core data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Streaming
kafka-python>=2.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Configuration and environment
python-dotenv
pydantic>=2.0.0

# Logging
structlog

# Date/time handling
python-dateutil

# Mathematical modeling
scikit-learn>=1.3.0

# Development and testing
pytest>=7.0.0
pytest-asyncio
